# docs/apispec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Contact Api",
    "version": "1.0",
    "description": "API SPEC for Contact Management",
    "contact": {
      "name": "Tubagus Dylan Rachmat",
      "email": "work.tubagusdylan@gmail.com"
    },
    "license": {
      "name": "MAVEN"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Server contact management API",
      "variables": {
        "environtment": {
          "default": "dev",
          "description": "Development Server"
        }
      }
    }
  ],
  "paths": {
...
    "/api/contacts": {
      "get": {
        "tags": ["Contact"],
        "summary": "Get search contact",
        "description": "Get list contact with search param",
        "parameters": [
          {
            "$ref": "#/components/parameters/HeaderToken"
          },
          {
            "name": "name",
            "description": "Search contact firstName or lastName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "phone",
            "description": "Search contact phone",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "email",
            "description": "Search contact email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "description": "Page for pagination",
            "required": true,
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "description": "Size per page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePaging"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 200,
                      "message": "OK",
                      "data": [
                        {
                          "id": "contact-laksjdlkasjd",
                          "firstName": "Udin",
                          "lastName": "Surudin",
                          "email": "udin@email.com",
                          "phone": "08969087878"
                        }
                      ],
                      "paging": {
                        "currentPage": 1,
                        "totalPage": 10,
                        "perPage": 10,
                        "totalData": 100
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "failed": {
                    "value": {
                      "code": 404,
                      "message": "Contact not found, ..."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Contact"],
        "summary": "Create new contact",
        "description": "Create new contact",
        "parameters": [
          {
            "$ref": "#/components/parameters/HeaderToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BodyContact"
              },
              "examples": {
                "udin": {
                  "value": {
                    "firstName": "Udin",
                    "lastName": "Surudin",
                    "email": "udin@email.com",
                    "phone": "08969087878"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 201,
                      "message": "OK",
                      "data": {
                        "id": "contact-laksjdlkasjd",
                        "firstName": "Udin",
                        "lastName": "Surudin",
                        "email": "udin@email.com",
                        "phone": "08969087878"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "failed": {
                    "value": {
                      "code": 400,
                      "message": "Email format invalid, ..."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/contacts/{contactId}": {
      "get": {
        "tags": ["Contact"],
        "summary": "Get contact",
        "description": "Get contact",
        "parameters": [
          {
            "$ref": "#/components/parameters/HeaderToken"
          },
          {
            "$ref": "#/components/parameters/ContactId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 200,
                      "message": "OK",
                      "data": {
                        "firstName": "ucup",
                        "lastName": "surucup",
                        "email": "ucup@email.com",
                        "phone": "0896121212"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "failed": {
                    "value": {
                      "code": 404,
                      "message": "Contact is not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Contact"],
        "summary": "Update contact",
        "description": "Update contact",
        "parameters": [
          {
            "$ref": "#/components/parameters/HeaderToken"
          },
          {
            "$ref": "#/components/parameters/ContactId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BodyContact"
              },
              "examples": {
                "udin": {
                  "value": {
                    "firstName": "ucup",
                    "lastName": "surucup",
                    "email": "ucup@email.com",
                    "phone": "0896121212"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 200,
                      "message": "OK",
                      "data": {
                        "firstName": "ucup",
                        "lastName": "surucup",
                        "email": "ucup@email.com",
                        "phone": "0896121212"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "failed": {
                    "value": {
                      "code": 400,
                      "message": "Email format invalid, ..."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Contact"],
        "summary": "Delete contact",
        "description": "Delete contact",
        "parameters": [
          {
            "$ref": "#/components/parameters/HeaderToken"
          },
          {
            "$ref": "#/components/parameters/ContactId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 200,
                      "message": "OK"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "failed": {
                    "value": {
                      "code": 404,
                      "message": "Contact is not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "HeaderToken": {
        "name": "X-API-TOKEN",
        "description": "Token for authorization",
        "required": true,
        "in": "header",
        "schema": {
          "type": "string"
        },
        "examples": {
          "token": {
            "description": "Sample token",
            "value": "asldaoichnalksjfoiqwoiehqwelkj"
          }
        }
      },
...
      "ContactId": {
        "name": "contactId",
        "description": "Contact id for get and update data",
        "required": true,
        "in": "path",
        "schema": {
          "type": "string"
        },
        "examples": {
          "user": {
            "description": "Sample contact id",
            "value": "contact-asldkjasld"
          }
        }
      },
...
    },
    "schemas": {
...
      "BodyContact": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ResponsePaging": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array"
          },
          "paging": {
            "type": "object"
          }
        }
      }
    }
  }
}